# api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_florinel-chis_gophercrm_internal_models.Customer:
    properties:
      address:
        type: string
      city:
        type: string
      company:
        type: string
      country:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      notes:
        type: string
      phone:
        type: string
      position:
        type: string
      postal_code:
        type: string
      state:
        type: string
      tickets:
        items:
          $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Ticket'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.User'
      user_id:
        type: integer
    type: object
...
  github_com_florinel-chis_gophercrm_internal_models.Ticket:
    properties:
      assigned_to:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.User'
      assigned_to_id:
        type: integer
      created_at:
        type: string
      customer:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Customer'
      customer_id:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      priority:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.TicketPriority'
      resolution:
        type: string
      status:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.TicketStatus'
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_florinel-chis_gophercrm_internal_models.TicketPriority:
    enum:
    - low
    - medium
    - high
    - urgent
    type: string
    x-enum-varnames:
    - TicketPriorityLow
    - TicketPriorityMedium
    - TicketPriorityHigh
    - TicketPriorityUrgent
  github_com_florinel-chis_gophercrm_internal_models.TicketStatus:
    enum:
    - open
    - in_progress
    - resolved
    - closed
    type: string
    x-enum-varnames:
    - TicketStatusOpen
    - TicketStatusInProgress
    - TicketStatusResolved
    - TicketStatusClosed
  github_com_florinel-chis_gophercrm_internal_models.User:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_login_at:
        type: string
      last_name:
        type: string
      locked_until:
        type: string
      role:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.UserRole'
      updated_at:
        type: string
    type: object
  github_com_florinel-chis_gophercrm_internal_models.UserRole:
    enum:
    - admin
    - sales
    - support
    - customer
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleSales
    - RoleSupport
    - RoleCustomer
  github_com_florinel-chis_gophercrm_internal_utils.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  github_com_florinel-chis_gophercrm_internal_utils.APIMeta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_florinel-chis_gophercrm_internal_utils.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
      meta:
        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIMeta'
      success:
        type: boolean
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handler.CreateCustomerRequest:
    properties:
      address:
        type: string
      city:
        type: string
      company:
        type: string
      country:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      notes:
        type: string
      phone:
        type: string
      position:
        type: string
      postal_code:
        type: string
      state:
        type: string
    required:
    - email
    - first_name
    - last_name
    type: object
...
info:
  contact: {}
  description: CRM API for managing users, leads, customers, tickets, tasks and API
    keys.
  title: GopherCRM API
  version: "1.0"
paths:
...
  /customers:
    get:
      description: List customers with pagination, optional search and sorting. Admin,
        sales and support roles see the same unfiltered set; the customer role is
        rejected outright and has no view of its own record here.
      parameters:
      - description: Page number (1-based); when supplied it overrides offset
        in: query
        name: page
        type: integer
      - default: 0
        description: Result offset, ignored when page is supplied
        in: query
        name: offset
        type: integer
      - default: 20
        description: Page size, capped at 100
        in: query
        name: limit
        type: integer
      - description: Sort column; ignored unless one of the allowed values
        enum:
        - created_at
        - updated_at
        - first_name
        - last_name
        - email
        - company
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort direction; anything else falls back to asc
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: Free-text search across customer fields; takes precedence over
          sort_by
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Customers retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                data:
                  properties:
                    customers:
                      items:
                        $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Customer'
                      type: array
                    total:
                      type: integer
                  type: object
                meta:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIMeta'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "403":
          description: Forbidden - Admin, Sales or Support role required
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "429":
          description: Too many requests - rate limit exceeded
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List customers
      tags:
      - customers
    post:
      consumes:
      - application/json
      description: Create a new customer (admin and sales roles only)
      parameters:
      - description: Customer creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Customer created successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_models.Customer'
              type: object
        "400":
          description: Invalid request data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "403":
          description: Forbidden - Admin or Sales role required
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "409":
          description: Customer with this email already exists
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "429":
          description: Too many requests - rate limit exceeded
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/github_com_florinel-chis_gophercrm_internal_utils.APIError'
              type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create a new customer
      tags:
      - customers
...
securityDefinitions:
  ApiKeyAuth:
    description: API key issued via POST /api-keys. The plaintext key is shown only
      once at creation.
    in: header
    name: ApiKey
    type: apiKey
  BearerAuth:
    description: 'JWT access token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"